requirements.txt
gradio
Pillow
numpy
opencv-python